backend/requirements.txt
flask
flask-cors
PyPDF2
langchain
langchain_google_genai
google-generativeai
faiss-cpu
python-dotenv
